Changed datatype for sizes to unsigned long ints

+++ src/functions.h
@@ -4,7 +4,7 @@ typedef struct {
   int hlinklens[5];
   char owner[128];
   char group[128];
-  int size[32];
+  unsigned long int size[32];
   int sizelens[32];
   char date[17];
   char name[512];
